+++ src/main/java/net/mcreator/thewetsmp/procedures/BambooJungleHillsFishProcedure.java
@@ -23,13 +23,13 @@ public class BambooJungleHillsFishProcedure {
 		double crab = 0;
 		double junglesnapper = 0;
 		/*fishing factors*/
-		luck = (double) LuckCheckProcedure.execute(entity);/*loot factors*/
-		time = (double) (world.dayTime());
-		altitude = (double) (Math.floor(y));/*define base values*/
-		crab = (double) 33;
-		junglesnapper = (double) 58;/*apply modifiers*/
+		luck = LuckCheckProcedure.execute(entity);/*loot factors*/
+		time = world.dayTime();
+		altitude = Math.floor(y);/*define base values*/
+		crab = 33;
+		junglesnapper = 58;/*apply modifiers*/
 		/*fish*/
-		fish = (double) junglesnapper;
+		fish = junglesnapper;
 		if (entity instanceof Player _player) {
 			Scoreboard _sc = _player.getScoreboard();
 			Objective _so = _sc.getObjective("PlayerRNG");
