+++ src/main/java/net/mcreator/thewetsmp/procedures/DeepColdOceanFishProcedure.java
@@ -26,30 +26,30 @@ public class DeepColdOceanFishProcedure {
 		double jellyfish = 0;
 		double luck = 0;
 		/*fishing factors*/
-		luck = (double) LuckCheckProcedure.execute(entity);/*loot factors*/
-		time = (double) (world.dayTime());
-		altitude = (double) (Math.floor(y));/*define base values*/
-		cod = (double) 10;
-		salmon = (double) 18;
-		scallop = (double) 23;
-		seaurchin = (double) 30;
-		jellyfish = (double) 38;/*apply modifiers*/
+		luck = LuckCheckProcedure.execute(entity);/*loot factors*/
+		time = world.dayTime();
+		altitude = Math.floor(y);/*define base values*/
+		cod = 10;
+		salmon = 18;
+		scallop = 23;
+		seaurchin = 30;
+		jellyfish = 38;/*apply modifiers*/
 		if (luck > 1) {
-			cod = (double) (cod - 3);
-			salmon = (double) (salmon - 2);
-			scallop = (double) (scallop + 5);
-			seaurchin = (double) (seaurchin + 4);
-			jellyfish = (double) (jellyfish + 2);
+			cod = cod - 3;
+			salmon = salmon - 2;
+			scallop = scallop + 5;
+			seaurchin = seaurchin + 4;
+			jellyfish = jellyfish + 2;
 		} else if (luck < 0) {
-			cod = (double) (cod + 3);
-			salmon = (double) (salmon + 2);
-			scallop = (double) (scallop - 2);
-			seaurchin = (double) (seaurchin - 4);
-			jellyfish = (double) (jellyfish - 3);
+			cod = cod + 3;
+			salmon = salmon + 2;
+			scallop = scallop - 2;
+			seaurchin = seaurchin - 4;
+			jellyfish = jellyfish - 3;
 		} /*fish*/
-		fish = (double) seaurchin;
+		fish = seaurchin;
 		if (time > 13000 && time < 23000) {
-			fish = (double) jellyfish;
+			fish = jellyfish;
 		}
 		if (entity instanceof Player _player) {
 			Scoreboard _sc = _player.getScoreboard();
