+++ src/main/java/net/mcreator/thewetsmp/procedures/DesertHillsFishProcedure.java
@@ -24,24 +24,24 @@ public class DesertHillsFishProcedure {
 		double glassfish = 0;
 		double crab = 0;
 		/*fishing factors*/
-		luck = (double) LuckCheckProcedure.execute(entity);/*loot factors*/
-		time = (double) (world.dayTime());
-		altitude = (double) (Math.floor(y));/*define base values*/
-		glassfish = (double) 5;
-		crab = (double) 16;
-		jellyfish = (double) 23;/*apply modifiers*/
+		luck = LuckCheckProcedure.execute(entity);/*loot factors*/
+		time = world.dayTime();
+		altitude = Math.floor(y);/*define base values*/
+		glassfish = 5;
+		crab = 16;
+		jellyfish = 23;/*apply modifiers*/
 		if (luck > 1) {
-			glassfish = (double) (glassfish + 8);
-			crab = (double) (crab - 1);
-			jellyfish = (double) (jellyfish + 3);
+			glassfish = glassfish + 8;
+			crab = crab - 1;
+			jellyfish = jellyfish + 3;
 		} else if (luck < 0) {
-			glassfish = (double) (glassfish - 2);
-			crab = (double) (crab + 4);
-			jellyfish = (double) (jellyfish + 2);
+			glassfish = glassfish - 2;
+			crab = crab + 4;
+			jellyfish = jellyfish + 2;
 		} /*fish*/
-		fish = (double) crab;
+		fish = crab;
 		if (altitude < 60) {
-			fish = (double) jellyfish;
+			fish = jellyfish;
 		}
 		if (entity instanceof Player _player) {
 			Scoreboard _sc = _player.getScoreboard();
